requirements.txt
opencc-python-reimplemented>=0.1.7
mutagen>=1.47.0
